backend/app/pyproject.toml
[tool.poetry]
name = "backend-template"
version = "0.1.0"
description = "Async fastapi+sqlalchemy(currently with sqlite) template for backend"
authors = ["Blinov Ilya"]
license = "CC"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.89.1"
pydantic = {extras = ["email"], version = "^1.10.7"}
SQLAlchemy = "^2.0.1"
asyncio = "^3.4.3"
aiosqlite = "^0.18.0"
httpx = "^0.23.3"
uvicorn = {extras = ["standard"], version = "^0.20.0"}
gunicorn = "^20.1.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
pyjwt = "^2.6.0"
fastapi-mail = "^1.2.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
